Updated paths and expanded some of the comments from Itsy's original code.
Commented out the Remove Wildlife variant function as it doesn't apply to most countries (only have teh data for mexico and brazil).
Had to input data in the main script instead.

Added an if file.exists for the time series (not all countries had data within the range) and some output vectors to print progress at the end.