Created (and updated) keys containing adm2 names and GID_2 codes from the GADM shape files.
Keys are per state per country.
Keys only exist so far for states which have surveillance data in SIRVERAdogs16.

The names are currently ascii and lower case. They should be remade as latin1 and ascii spelling added only when they are exceptions, as some placenames are differentiated by their non-ascii characters. (e.g in BRA.10_1 this was an issue).